Eliminate the CPU-side copy in `StorageBuffer`.
Today, when we upload data to a `StorageBuffer`, it must be copied
twice: once from `StorageBuffer::value` to `StorageBuffer::scratch`, and
once from `StorageBuffer::scratch` to the GPU. This patch eliminates the
former copy in favor of storing the single canonical copy of the data
directly inside the `scratch` field.

This patch complements #12489, which eliminates both copies for mesh
data at the cost of requiring more overhead for GPU I/O, especially for
single-threaded situations. Simply eliminating one of the copies seems
an unambiguous win (and is lower-hanging fruit), so both this patch and
PR #12489 are valuable.