Implement sheen per the `KHR_materials_sheen` glTF extension.
Like clearcoat, sheen is an extra PBR layer on top of the base layer.
It's designed to represent cloth and fabrics, which the standard PBR
model often fails to capture well due to forward and backward
scattering. See [the Filament spec] for more details, including visual
comparisons.

A new example, `sheen`, has been added. It displays the example sheen
chair from the [glTF Sample Assets] repository. To reduce the size of
the files in the Bevy repository, the textures have been shrunk and
recompressed.

[the Filament spec]: https://google.github.io/filament/Materials.md.html#materialmodels/clothmodel

[glTF Sample Assets]: https://github.com/KhronosGroup/glTF-Sample-Assets