Allow animation clips to animate arbitrary reflectable properties.
Currently, Bevy restricts animation clips to animating
`Transform::translation`, `Transform::rotation`, `Transform::scale`, or
`MorphWeights`, which correspond to the properties that glTF can
animate. This is insufficient for many use cases such as animating UI,
as the UI layout systems expect to have exclusive control over UI
elements' `Transform`s and therefore the `Style` properties must be
animated instead.

To fix this problem, cart proposed in Discord that Bevy use the
reflection system in order to animate arbitrary properties of
components, as long as those properties are `Animatable`.

This commit implements that approach, allowing for `AnimationClip`s to
name arbitrary properties to be animated. The `Keyframes` structure has
been reworked (and moved into its own module) to take the type ID of a
component, a path from the component to the property to be animated, and
a dynamically-typed list of keyframes. Convenience methods have been
added to `Keyframes` to replicate the old APIs for the special cases of
translation, rotation, scale, and morph weights. For the most part, you
can migrate by simply changing your code from
`Keyframes::Translation(...)` to `Keyframes::translation(...)`.

Because the `AnimationPlayer` and `AnimationTarget` components are
extremely flexible and can coexist anywhere in the entity hierarchy, and
because animations can now legally animate any component, no static
query suffices to enable parallelism over `AnimationTarget`s and their
components to be animated while simultaneously allowing for read-only
access to `AnimationPlayer`s anywhere in the hierarchy. (The reason why
`With` and `Without` can't be used is that entities can legally have
both `AnimationPlayer` and `AnimationTarget` components--furthermore, an
entity containing an `AnimationPlayer` and an `AnimationTarget` can
legally be targeted by *another* entity containing an `AnimationPlayer`
and an `AnimationTarget`--and we rely on this in at least one example.)
Therefore, we must dynamically construct queries using the
`SystemParamBuilder` API, surfacing the components to be mutated as a
`FilteredEntityMut`.

`FilteredEntityMut` hasn't currently been optimized for performance, so
by itself, this patch significantly regresses performance on
`many_foxes`. Separately, I'm preparing a PR that fixes the performance
problems with `FilteredEntityMut`. With that PR applied on top of this
one, I've measured a performance regression of around 18% on
`many_foxes`. This is unfortunate, but I believe it's worth it for the
significantly added flexibility that this patch provides.

A new example, `animated_ui`, has been added. This example shows how to
use Bevy's built-in animation infrastructure to animate font size and
color, which wasn't possible before this patch.