Put `FilteredEntityRef::access` and `FilteredEntityMut::access` behind `Arc`s.
Currently, `FilteredEntityRef` and `FilteredEntityMut` clone `access`
whenever they're cloned, which happens frequently with fetches and with
methods like `reborrow()`. `access` contains a multitude of
`FixedBitSet`s, which add a lot of allocation traffic. The forthcoming
reflection-based animation PR heavily uses `FilteredEntityMut` in
performance-sensitive code, and cloning all those bitsets was taking
30%-40% of the time.

It was suggested in Discord to place the `access` field behind an `Arc`
to avoid all the memory traffic on clone, and this is what this patch
does. It significantly improves performance of animation-heavy workloads
like `many_foxes` when reflection-based animation is in use.