Implement basic clustered decal projectors.
This commit adds support for *decal projectors* to Bevy, allowing for
textures to be projected on top of geometry. Decal projectors are
clusterable objects, just as punctual lights and light probes are. This
means that decals are only evaluated for objects within the conservative
bounds of the projector, and they don't require a second pass.

There are several limitations of this initial patch that can be
addressed in follow-ups:

1. There's no way to specify the Z-index of decals. That is, the order
   in which multiple decals are blended on top of one another is
   arbitrary. A follow-up could introduce some sort of Z-index field so
   that artists can specify that some decals should be blended on top of
   others.

2. Decals don't take the normal of the surface they're projected onto
   into account. Most decal implementations in other engines have a
   feature whereby the angle between the decal projector and the normal
   of the surface must be within some threshold for the decal to appear.
   Often, artists can specify a fade-off range for a smooth transition
   between oblique surfaces and aligned surfaces.

3. There's no distance-based fadeoff toward the end of the projector
   range. Many decal implementations have this.